Augment functor interface for uprase_fn and upsert.
It is often useful to be able to operate over a newly-inserted value in the
map, just as it is to operate over an already-existing value.

To enable this functionality, we support an additional overload to the functor
passed to uprase_fn or upsert, which implements the signature
operator()(mapped_type&, UpsertContext), where UpsertContext is an enum
describing whether the value was newly-inserted or already existed in the
table.

To mantain backwards compatibility, if the user's functor only implements
operator()(mapped_type&), we will only invoke it if the value already existed
in the table.